requirements.txt
fastapi==0.100.0
uvicorn==0.20.0
pymongo==4.4.0
email-validator==2.0.0
pydantic-settings==2.0.2
casbin==1.22.0
casbin_pymongo_adapter==1.0.0
passlib==1.7.4
PyJWT==2.8.0
python-multipart==0.0.6
orjson==3.9.4
pyyaml==6.0.1